fix: 출석부 목록, 명단 수정 시 요일 정렬되도록 수정
